+++ CHANGELOG.md
@@ -3,10 +3,11 @@ This changelog is inspired by [keepachangelog.com](https://keepachangelog.com/en
 
 ## [0.1.2.0] - 2022-02-19
 ### Added
- - 12 new voting methods:
+ - 14 new voting methods:
     - positional: borda, nauru, eurovision, dabagh, binary_positional
     - evaluative: approval, disapproval, cav, score, range
     - condorcet: copeland, lull
+    - cumulative: cumulative, fractional
  - styling: amazing scroll shadows
 
 ### Improved